requirements.txt
langchain-text-splitters
langchain-core
# Core dependencies
langchain>=0.1.0
langchain-community>=0.0.20
langchain-openai>=0.0.5
langchain-chroma>=0.1.0

# Vector store
chromadb>=0.4.22

# OpenAI
openai>=1.10.0

# Embeddings
langchain-huggingface

# UI
gradio>=4.16.0

# Environment variables
python-dotenv>=1.0.0

# Ollama support (optional)
langchain-ollama>=0.0.1

# Utilities
typing-extensions>=4.9.0
